refactor events collection
