+++ teamcity/assets/configuration/spec.yaml
@@ -7,19 +7,6 @@ files:
     - template: init_config/default
   - template: instances
     options:
-    - name: bearer_token
-      description: Bearer token for HTTP Token Authentication to the TeamCity Server's REST API.
-      value:
-        type: string
-        example: <BEARER_TOKEN>
-      display_priority: 1
-    - name: name
-      description: |
-        A custom unique name per build configuration that is shown
-        in the events.
-      value:
-        type: string
-        example: <BUILD_NAME>
     - name: server
       description: |
         Specify the server name of your teamcity instance
@@ -30,6 +17,31 @@ files:
       value:
         type: string
         example: http://teamcity.<ACCOUNT_NAME>.com
+    - name: use_openmetrics
+      description: |
+        Set to true to enable OpenMetric collection from the TeamCity server's metrics endpoint.
+      value:
+        type: boolean
+        example: false
+    - name: experimental_metrics
+      description: |
+        Set to true to enable collection of experimental metrics.
+      value:
+        type: boolean
+        example: true
+    - name: collect_events
+      description: |
+        Set to true to enable collection of TeamCity build and deployment success and failure events
+      value:
+        type: boolean
+        example: true
+    - name: name
+      description: |
+        A custom unique name per build configuration that is shown
+        in the events.
+      value:
+        type: string
+        example: <BUILD_NAME>
     - name: build_configuration
       description: |
         This is the internal build ID of the build configuration you wish to track.
@@ -57,10 +69,8 @@ files:
         example: false
     - template: instances/openmetrics
       overrides:
-        openmetrics_endpoint.value.example: http://localhost:8111/app/metrics
-        openmetrics_endpoint.display_priority: 2
+        openmetrics_endpoint.hidden: true
         openmetrics_endpoint.required: false
-        openmetrics_endpoint.enabled: true
   - template: logs
     example:
     - type: file
