+++ teamcity/datadog_checks/teamcity/config_models/instance.py
@@ -81,11 +81,11 @@ class InstanceConfig(BaseModel):
     aws_region: Optional[str]
     aws_service: Optional[str]
     basic_http_authentication: Optional[bool]
-    bearer_token: Optional[str]
     build_configuration: Optional[str]
     cache_metric_wildcards: Optional[bool]
     cache_shared_labels: Optional[bool]
     collect_counters_with_distributions: Optional[bool]
+    collect_events: Optional[bool]
     collect_histogram_buckets: Optional[bool]
     connect_timeout: Optional[float]
     disable_generic_tags: Optional[bool]
@@ -94,6 +94,7 @@ class InstanceConfig(BaseModel):
     exclude_labels: Optional[Sequence[str]]
     exclude_metrics: Optional[Sequence[str]]
     exclude_metrics_by_labels: Optional[Mapping[str, Union[bool, Sequence[str]]]]
+    experimental_metrics: Optional[bool]
     extra_headers: Optional[Mapping[str, Any]]
     extra_metrics: Optional[Sequence[Union[str, Mapping[str, Union[str, ExtraMetric]]]]]
     headers: Optional[Mapping[str, Any]]
@@ -145,6 +146,7 @@ class InstanceConfig(BaseModel):
     tls_verify: Optional[bool]
     use_latest_spec: Optional[bool]
     use_legacy_auth_encoding: Optional[bool]
+    use_openmetrics: Optional[bool]
     use_process_start_time: Optional[bool]
     username: Optional[str]
